requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.22
djangorestframework==3.15.2
dotenv==0.9.9
python-dotenv==1.0.1
pytz==2025.2
sqlparse==0.5.3
typing-extensions==4.13.2
